requirements.txt
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.6
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.9.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
